Maths becomes law-proved, World decomposes, and Post stops being a gate
The trunk since 3cfe27fb, landed whole. What follows is what changed and what
was actually run against the tree this commit installs.

Puck.Maths is now proved by law rather than by assertion. tests/Puck.Maths.Tests
is a declaration-first law suite over the library's public surface: fixed point
(signed and unsigned Q48.16, unit fractions, vectors, kinematics, hypercomplex),
finite fields (binary polynomials, binary, prime and extension fields), and the
sampling wing. An adversarial excursion ran against the whole library and its
findings closed with regressions that reach the input that broke them: parsing
and formatting at culture boundaries, Pow over negative bases, CreateTruncating's
three modes, zero-state policy for the three field descriptors, factorization
above the witness boundary. Every wing carries a README written for people and a
doc-gap register that closes on the code rather than on the prose.

Verified code can no longer drift unnoticed. Puck.Analyzers enforces the
[VerifiedCode] brand: VER001 through VER010 seal a branded declaration together
with the dependencies its entry names, refuse a ledger that cannot be trusted,
and refuse a brand that nothing can record. Twenty-five bypasses that held the
brand open are shut.

Puck.World decomposes into Puck.World.Data, Puck.World.Server, and a composition
root. The document (puck.world.def.v1) describes itself and refuses unmapped
members on every row. One ordered submission domain carries commands, typed
completions, and the output hub; the HUD is document state; game state is
document rows the engine never interprets. Authority stops trusting the client:
every mutating ingress consults the acting principal before mutation, a decision
is data with a readable refusal, and grants name their verbs. The capability
channel campaign ran to its end -- folds, consent as a player-owned axis, and
co-driving where a human and an addon reach the same body.

The charter narrowed to four worlds. Play is the hub; Dive, Kart, and Jump are
reached through picture-frame portals rather than a flag or a menu. Motion is
authored data, not an engine opinion.

The tree got smaller where it should have. experimental/ takes Puck.Post,
Puck.Demo, Puck.Bench, the tools and scripts trees, and Puck.BareMetal;
Puck.Scene, Puck.Cameras, Puck.Capture, Puck.Authoring and Puck.Compositing
cease to exist, merged into the projects that already owned their ideas. The
capability catalog and register are deleted rather than corrected: both claimed
per-capability verification status that no longer existed. Puck.Cli grows into
the developer CLI and carries the checks that survived -- search, references,
declarations, architecture, format, citations, landing.

Proven against this exact tree:

  dotnet build Puck.slnx -c Release      0 warnings, 0 errors, 36 projects.
                                         This is where the architecture gate and
                                         the brand enforcer actually run.
  dotnet test Puck.slnx -c Release       632 passed, 0 failed
                                         (Maths 471, Analyzers 145, World 16).
  puck citations                         exit 0; 243 verbs enumerated, 171
                                         registered, 408 verb-shaped literals,
                                         0 unresolved.
  puck architecture                      exit 0; every declared check passed.
  Puck.World, --exit-after-seconds 2     Window opened 1920x1080, VRR pacing
                                         negotiated, seat paired, clean
                                         shutdown, exit 0.
  Full Release rebuild                   Left the tree clean, so every committed
                                         .spv and .dxil is what its HLSL
                                         compiles to.

How these proofs can fail:

Puck.World.Tests is sixteen cases -- a proof of shape, not coverage of the
world. The Maths suite leaves 634 of 2120 public members uncovered and 10
waived; a green run says the covered laws held, not that the library is right.
The world boot proves the engine reaches a presented frame on this machine's
AMD FreeSync path and nothing about what the frame contained. puck citations
proves cited names resolve, never that the prose around them is true.

What has no gate at all: Puck.Post is quarantined, so the shared engine contract
it used to cover -- cross-backend render parity, the SDF VM ISA, document
schemas, deterministic numerics -- is unguarded today. That is a live gap, not
a covered one, and no battery in this landing replaces it.

+++ src/Puck.Text/TextEffectParameter.cs
@@ -11,7 +11,7 @@ namespace Puck.Text;
 /// <param name="BindingMode">How the bound variable folds onto <see cref="BaseValue"/>.</param>
 public readonly record struct TextEffectParameter(
     float BaseValue,
-    uint VariableHash = 0,
+    ulong VariableHash = 0,
     TextEffectParameterBindingMode BindingMode = TextEffectParameterBindingMode.Multiplicative
 ) {
     /// <summary>Resolves the parameter against the current content-time channels.</summary>
